Generate Lorem Ipsum using the Loripsum API

Generates random, customisable placeholder text using the Loripsum Web API.

Usage

loripsum(
  n = 5L,
  length = "short",
  decorate = FALSE,
  link = FALSE,
  ul = FALSE,
  ol = FALSE,
  dl = FALSE,
  bq = FALSE,
  code = FALSE,
  headers = FALSE,
  allcaps = FALSE,
  prude = FALSE,
  plaintext = FALSE
)

Source

https://loripsum.net/

Arguments

n

Integer. The number of paragraphs to generate. Defaults to 5.

length

Character. The average length of a paragraph. Can be any of "short", "medium", "long" or "verylong". Defaults to "short".

decorate

Logical. If TRUE, bold, italic and marked text. Defaults to FALSE.

link

Logical. If TRUE, adds links. Defaults to FALSE.

ul

Logical. If TRUE, adds unordered lists. Defaults to FALSE.

ol

Logical. If TRUE, adds numbered/ordered lists. Defaults to FALSE.

dl

Logical. If TRUE, adds description lists. Defaults to FALSE.

bq

Logical. If TRUE, adds blockquotes. Defaults to FALSE.

code

Logical. If TRUE, adds code samples. Defaults to FALSE.

headers

Logical. If TRUE, adds headers. Defaults to FALSE.

allcaps

Logical. If TRUE, uses ALL CAPS. Defaults to FALSE.

prude

Logical. If TRUE, removes Latin words which could be construed as offensive in English, like 'sex' or 'homo'. Defaults to FALSE.

plaintext

Logical. If TRUE, returns plain text without HTML tags. Defaults to FALSE.

See also